backend/requirements.txt
Django
djangorestframework==3.14.0
gunicorn==21.2.0
django-cors-headers==4.4.0
